fix(aco): corrige maxPenalty, evaporação/seleção e remove GAMA
Elimina confusão de tipos em maxPenalty, ajusta evaporação e seleção de
formigas, remove o parâmetro GAMA e o default de rho passa a 0,2.

+++ src/graph/math.go
@@ -25,13 +25,10 @@ func angle(v1, v2 vector2D) float64 {
 	return math.Acos(cosTheta)
 }
 
-//BUG: maxPenalti como time.Duration (int64) — float64(maxPenalti) só não dá overflow porque 1ns→1.0.
-//     O TODO original reportava valores ~5e+08 no grafo, provavelmente NaN propagation ou bug de versão anterior.
-//     Se maxPenalti fosse alterado para 1e9, o cast para nanossegundos daria 0, subestimando o turn cost.
 func turnCost(v1, v2 vector2D) float64 {
 	angle := angle(v1, v2)
 	if math.IsNaN(angle) {
 		log.Printf("NaN detected in angle calculation: v1=%v, v2=%v", v1, v2)
 	}
-	return float64(maxPenalti) * (angle / math.Pi)
+	return maxPenalti * (angle / math.Pi)
 }
